regulator: tps62361: Fix build Warnings
Fix warnings:
-cast from pointer to integer of different size, 
 in drivers/regulator/tps6586x-regulator.c file.
 
Bug 1566945

Reviewed-on: http://git-master/r/592698
GVS: Gerrit_Virtual_Submit

+++ drivers/regulator/tps62360-regulator.c
@@ -354,7 +354,7 @@ static int tps62360_probe(struct i2c_client *client,
 	struct tps62360_chip *tps;
 	int ret;
 	int i;
-	int chip_id;
+	uintptr_t chip_id;
 
 	pdata = client->dev.platform_data;
 	chip_id = id->driver_data;
@@ -367,7 +367,7 @@ static int tps62360_probe(struct i2c_client *client,
 			dev_err(&client->dev, "Error: No device match found\n");
 			return -ENODEV;
 		}
-		chip_id = (int)match->data;
+		chip_id = (uintptr_t)match->data;
 		if (!pdata)
 			pdata = of_get_tps62360_platform_data(&client->dev);
 	}
